+++ common/src/main/java/nl/enjarai/doabarrelroll/DoABarrelRollClient.java
@@ -21,8 +21,6 @@ import nl.enjarai.doabarrelroll.util.MixinHooks;
 import nl.enjarai.doabarrelroll.util.Vec2d;
 
 public class DoABarrelRollClient {
-    public static final String MODID = "do_a_barrel_roll";
-
     public static final SmoothUtil pitchSmoother = new SmoothUtil();
     public static final SmoothUtil yawSmoother = new SmoothUtil();
     public static final SmoothUtil rollSmoother = new SmoothUtil();
@@ -34,10 +32,6 @@ public class DoABarrelRollClient {
     public static double throttle = 0;
 
     // TODO triple jump to activate???
-
-    public static Identifier id(String path) {
-        return new Identifier(MODID, path);
-    }
 
 
     public static void clientTick(MinecraftClient client) {