+++ forge/src/main/java/nl/enjarai/doabarrelroll/forge/ForgeModBusEvents.java
@@ -1,9 +1,0 @@
-package nl.enjarai.doabarrelroll.forge;
-
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.fml.common.Mod;
-
-@Mod.EventBusSubscriber(value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
-public class ForgeModBusEvents {
-
-}